Preserves parent context in `{{#flipper}}` block
A block helper implicitly creates a new data context. This would cause
the parent context to get lost within a `{{#flipper}}` block, which is
not great. The whole point is to be able to wrap the old/existing
behaviour within an `{{else}}...{{/flipper}}` pair, and this should run
as if the flipper code didn't exist at all when the flipper is off. Not
so without also passing the expected data context along to the
`Template.elseBlock`. Fixes #1.