Fixes no-account bug
If we have a separate service travel through a flippered code path, or
mistakenly use it within a publish function, it blows up on the internal
`Meteor.userId()` check, since `Accounts` isn't really available. This
is a kind of crap side effect that relies on the caller and the flipper
check should clean up after itself, especially here where the answer to
there being no user is obvious: drop back to boolean gates only.

This is a better solution to:

https://github.com/stage3systems/charter-watch/pull/164
https://github.com/stage3systems/charter-watch/pull/165

The place where Meteor blows up is:

https://github.com/meteor/meteor/blob/master/packages/accounts-base/accounts_server.js#L80